udf{info,label,tune}: Migrate common code into {read,update}disc.[ch]
