+++ udfinfo/main.c
@@ -38,52 +38,6 @@
 #include "options.h"
 #include "readdisc.h"
 
-static uint64_t get_size(int fd)
-{
-	struct stat st;
-	uint64_t size;
-	off_t offset;
-
-	if (fstat(fd, &st) == 0)
-	{
-		if (S_ISBLK(st.st_mode) && ioctl(fd, BLKGETSIZE64, &size) == 0)
-			return size;
-		else if (S_ISREG(st.st_mode))
-			return st.st_size;
-	}
-
-	offset = lseek(fd, 0, SEEK_END);
-	if (offset == (off_t)-1)
-	{
-		fprintf(stderr, "%s: Error: Cannot detect size of disk: %s\n", appname, strerror(errno));
-		exit(1);
-	}
-
-	if (lseek(fd, 0, SEEK_SET) != 0)
-	{
-		fprintf(stderr, "%s: Error: Cannot seek to start of disk: %s\n", appname, strerror(errno));
-		exit(1);
-	}
-
-	return offset;
-}
-
-static int get_sector_size(int fd)
-{
-	int size;
-
-	if (ioctl(fd, BLKSSZGET, &size) != 0)
-		return 0;
-
-	if (size < 512 || size > 32768 || (size & (size - 1)))
-	{
-		fprintf(stderr, "%s: Warning: Disk logical sector size (%d) is not suitable for UDF\n", appname, size);
-		return 0;
-	}
-
-	return size;
-}
-
 static uint32_t compute_windows_serial_num(struct udf_disc *disc)
 {
 	uint8_t *buffer = (uint8_t *)disc->udf_fsd;
