+++ udflabel/Makefile.am
@@ -1,4 +1,4 @@
 sbin_PROGRAMS = udflabel
 udflabel_LDADD = $(top_builddir)/libudffs/libudffs.la
-udflabel_SOURCES = main.c options.c ../udfinfo/readdisc.c options.h ../udfinfo/readdisc.h ../include/ecma_167.h ../include/osta_udf.h ../include/libudffs.h ../include/bswap.h
+udflabel_SOURCES = main.c options.c ../udfinfo/readdisc.c ../udftune/updatedisc.c options.h ../udftune/updatedisc.h ../udfinfo/readdisc.h ../include/ecma_167.h ../include/osta_udf.h ../include/libudffs.h ../include/bswap.h
 AM_CPPFLAGS = -I$(top_srcdir)/include
