+++ udftune/Makefile.am
@@ -1,4 +1,4 @@
 sbin_PROGRAMS = udftune
 udftune_LDADD = $(top_builddir)/libudffs/libudffs.la
-udftune_SOURCES = main.c options.c ../udfinfo/readdisc.c options.h ../udfinfo/readdisc.h ../include/ecma_167.h ../include/osta_udf.h ../include/libudffs.h
+udftune_SOURCES = main.c options.c updatedisc.c ../udfinfo/readdisc.c updatedisc.h options.h ../udfinfo/readdisc.h ../include/ecma_167.h ../include/osta_udf.h ../include/libudffs.h
 AM_CPPFLAGS = -I$(top_srcdir)/include
